app/requirements.txt
fastapi
uvicorn
langchain
langgraph
langchain-community
langchain-openai
faiss-cpu
openai
pydantic
python-dotenv
tiktoken
